Split iOS progress badge refresh tests
