+++ apps/ios/Flashcards/FlashcardsTests/Progress/Refresh/ProgressReviewBadgeRefreshTests.swift
@@ -1,0 +1,341 @@
+import Foundation
+import XCTest
+@testable import Flashcards
+
+final class ProgressReviewBadgeRefreshTests: ProgressStoreTestCase {
+    @MainActor
+    func testRefreshReviewProgressBadgeKeepsLocalTodayReviewWhenServerSummaryIsStale() async throws {
+        let database = try self.makeDatabase()
+        let workspace = try database.workspaceSettingsStore.loadWorkspace()
+        let cloudSettings = try database.workspaceSettingsStore.loadCloudSettings()
+
+        let now = try XCTUnwrap(parseIsoTimestamp(value: "2026-04-18T12:00:00.000Z"))
+        let timeZone = try XCTUnwrap(TimeZone(identifier: "UTC"))
+        let requestRange = try makeTestProgressRequestRange(
+            now: now,
+            timeZone: timeZone,
+            dayCount: 140
+        )
+        try self.addReviewedCard(
+            database: database,
+            workspaceId: workspace.workspaceId,
+            reviewedAtClient: try makeReviewedAtClientForTests(
+                localDate: "2026-04-18",
+                hour: 9,
+                timeZoneIdentifier: requestRange.timeZone
+            )
+        )
+        let outboxEntries = try database.loadOutboxEntries(workspaceId: workspace.workspaceId, limit: Int.max)
+        try database.deleteOutboxEntries(operationIds: outboxEntries.map(\.operationId))
+
+        let serverSeries = try makeTestProgressSeries(
+            requestRange: requestRange,
+            reviewCountsByDate: [:],
+            generatedAt: "2026-04-18T11:59:00.000Z"
+        )
+        let serverSummary = try makeTestProgressSummary(
+            timeZone: requestRange.timeZone,
+            reviewDates: [],
+            generatedAt: "2026-04-18T11:59:00.000Z"
+        )
+        let context = try self.makeProgressStoreContext(
+            database: database,
+            workspaceId: workspace.workspaceId,
+            installationId: cloudSettings.installationId,
+            serverSummary: serverSummary,
+            serverSeries: serverSeries,
+            loadProgressSummaryError: nil,
+            loadProgressSeriesError: nil,
+            cloudState: .guest
+        )
+        defer { context.tearDown() }
+
+        await context.store.refreshReviewProgressBadgeIfNeeded(now: now)
+
+        XCTAssertNil(context.store.progressSnapshot)
+        XCTAssertEqual(
+            ReviewProgressBadgeState(
+                streakDays: 1,
+                hasReviewedToday: true,
+                streakFreezeAvailableCredits: 2,
+                streakFreezeCapacity: 2,
+                showsStreakFreezeBank: true,
+                isInteractive: true
+            ),
+            context.store.reviewProgressBadgeState
+        )
+        XCTAssertEqual(1, context.cloudSyncService.loadProgressSummaryCallCount)
+        XCTAssertEqual(0, context.cloudSyncService.loadProgressSeriesCallCount)
+    }
+
+    @MainActor
+    func testRefreshReviewProgressBadgePatchesServerBaseWithPendingTodayReview() async throws {
+        let database = try self.makeDatabase()
+        let workspace = try database.workspaceSettingsStore.loadWorkspace()
+        let cloudSettings = try database.workspaceSettingsStore.loadCloudSettings()
+        let timeZone = try XCTUnwrap(TimeZone(identifier: "UTC"))
+
+        let now = try XCTUnwrap(parseIsoTimestamp(value: "2026-04-18T12:00:00.000Z"))
+        let requestRange = try makeTestProgressRequestRange(
+            now: now,
+            timeZone: timeZone,
+            dayCount: 140
+        )
+        try self.addReviewedCard(
+            database: database,
+            workspaceId: workspace.workspaceId,
+            reviewedAtClient: try makeReviewedAtClientForTests(
+                localDate: "2026-04-18",
+                hour: 9,
+                timeZoneIdentifier: requestRange.timeZone
+            )
+        )
+
+        let serverSeries = try makeTestProgressSeries(
+            requestRange: requestRange,
+            reviewCountsByDate: [
+                "2026-04-17": 1
+            ],
+            generatedAt: "2026-04-18T11:59:00.000Z"
+        )
+        let serverSummary = UserProgressSummary(
+            timeZone: requestRange.timeZone,
+            summary: makeTestProgressSummaryValue(
+                currentStreakDays: 200,
+                hasReviewedToday: false,
+                lastReviewedOn: "2026-04-17",
+                activeReviewDays: 200
+            ),
+            generatedAt: "2026-04-18T11:59:00.000Z",
+            reviewHistoryWatermarks: makeTestProgressReviewHistoryWatermarks(reviewSequenceId: 42)
+        )
+        let context = try self.makeProgressStoreContext(
+            database: database,
+            workspaceId: workspace.workspaceId,
+            installationId: cloudSettings.installationId,
+            serverSummary: serverSummary,
+            serverSeries: serverSeries,
+            loadProgressSummaryError: nil,
+            loadProgressSeriesError: nil,
+            cloudState: .guest
+        )
+        defer { context.tearDown() }
+
+        await context.store.refreshReviewProgressBadgeIfNeeded(now: now)
+
+        XCTAssertNil(context.store.progressSnapshot)
+        XCTAssertEqual(
+            ReviewProgressBadgeState(
+                streakDays: 201,
+                hasReviewedToday: true,
+                streakFreezeAvailableCredits: 2,
+                streakFreezeCapacity: 2,
+                showsStreakFreezeBank: true,
+                isInteractive: true
+            ),
+            context.store.reviewProgressBadgeState
+        )
+        XCTAssertEqual(1, context.cloudSyncService.loadProgressSummaryCallCount)
+        XCTAssertEqual(0, context.cloudSyncService.loadProgressSeriesCallCount)
+    }
+
+    @MainActor
+    func testRefreshReviewProgressBadgeExtendsLongServerStreakWithLocalTodayReview() async throws {
+        let database = try self.makeDatabase()
+        let workspace = try database.workspaceSettingsStore.loadWorkspace()
+        let cloudSettings = try database.workspaceSettingsStore.loadCloudSettings()
+        let timeZone = try XCTUnwrap(TimeZone(identifier: "UTC"))
+        try self.addReviewedCard(
+            database: database,
+            workspaceId: workspace.workspaceId,
+            reviewedAtClient: try makeReviewedAtClientForTests(
+                localDate: "2026-04-18",
+                hour: 9,
+                timeZoneIdentifier: timeZone.identifier
+            )
+        )
+        let outboxEntries = try database.loadOutboxEntries(workspaceId: workspace.workspaceId, limit: Int.max)
+        try database.deleteOutboxEntries(operationIds: outboxEntries.map(\.operationId))
+
+        let now = try XCTUnwrap(parseIsoTimestamp(value: "2026-04-18T12:00:00.000Z"))
+        let requestRange = try makeTestProgressRequestRange(
+            now: now,
+            timeZone: timeZone,
+            dayCount: 140
+        )
+        let serverSeries = try makeTestProgressSeries(
+            requestRange: requestRange,
+            reviewCountsByDate: [
+                "2026-04-17": 1
+            ],
+            generatedAt: "2026-04-18T11:59:00.000Z"
+        )
+        let serverSummary = UserProgressSummary(
+            timeZone: requestRange.timeZone,
+            summary: makeTestProgressSummaryValue(
+                currentStreakDays: 200,
+                hasReviewedToday: false,
+                lastReviewedOn: "2026-04-17",
+                activeReviewDays: 200
+            ),
+            generatedAt: "2026-04-18T11:59:00.000Z",
+            reviewHistoryWatermarks: makeTestProgressReviewHistoryWatermarks(reviewSequenceId: 42)
+        )
+        let context = try self.makeProgressStoreContext(
+            database: database,
+            workspaceId: workspace.workspaceId,
+            installationId: cloudSettings.installationId,
+            serverSummary: serverSummary,
+            serverSeries: serverSeries,
+            loadProgressSummaryError: nil,
+            loadProgressSeriesError: nil,
+            cloudState: .guest
+        )
+        defer { context.tearDown() }
+
+        await context.store.refreshReviewProgressBadgeIfNeeded(now: now)
+
+        XCTAssertNil(context.store.progressSnapshot)
+        XCTAssertEqual(
+            ReviewProgressBadgeState(
+                streakDays: 201,
+                hasReviewedToday: true,
+                streakFreezeAvailableCredits: 2,
+                streakFreezeCapacity: 2,
+                showsStreakFreezeBank: true,
+                isInteractive: true
+            ),
+            context.store.reviewProgressBadgeState
+        )
+        XCTAssertEqual(1, context.cloudSyncService.loadProgressSummaryCallCount)
+        XCTAssertEqual(0, context.cloudSyncService.loadProgressSeriesCallCount)
+    }
+
+    @MainActor
+    func testRefreshReviewProgressBadgeDoesNotExtendServerFrozenReplacementWithoutSeriesBase() async throws {
+        let database = try self.makeDatabase()
+        let workspace = try database.workspaceSettingsStore.loadWorkspace()
+        let cloudSettings = try database.workspaceSettingsStore.loadCloudSettings()
+        let timeZone = try XCTUnwrap(TimeZone(identifier: "UTC"))
+        try self.addReviewedCard(
+            database: database,
+            workspaceId: workspace.workspaceId,
+            reviewedAtClient: try makeReviewedAtClientForTests(
+                localDate: "2026-04-17",
+                hour: 9,
+                timeZoneIdentifier: timeZone.identifier
+            )
+        )
+        let outboxEntries = try database.loadOutboxEntries(workspaceId: workspace.workspaceId, limit: Int.max)
+        try database.deleteOutboxEntries(operationIds: outboxEntries.map(\.operationId))
+
+        let now = try XCTUnwrap(parseIsoTimestamp(value: "2026-04-18T12:00:00.000Z"))
+        let requestRange = try makeTestProgressRequestRange(
+            now: now,
+            timeZone: timeZone,
+            dayCount: 140
+        )
+        let serverSeries = try makeTestProgressSeries(
+            requestRange: requestRange,
+            reviewCountsByDate: [
+                "2026-04-16": 1
+            ],
+            generatedAt: "2026-04-18T11:59:00.000Z"
+        )
+        let serverSummary = UserProgressSummary(
+            timeZone: requestRange.timeZone,
+            summary: ProgressSummary(
+                currentStreakDays: 2,
+                longestStreakDays: 2,
+                hasReviewedToday: false,
+                lastReviewedOn: "2026-04-16",
+                activeReviewDays: 1,
+                streakFreeze: makeTestProgressStreakFreeze(availableCredits: 1, balanceUnits: 11)
+            ),
+            generatedAt: "2026-04-18T11:59:00.000Z",
+            reviewHistoryWatermarks: makeTestProgressReviewHistoryWatermarks(reviewSequenceId: 42)
+        )
+        let context = try self.makeProgressStoreContext(
+            database: database,
+            workspaceId: workspace.workspaceId,
+            installationId: cloudSettings.installationId,
+            serverSummary: serverSummary,
+            serverSeries: serverSeries,
+            loadProgressSummaryError: nil,
+            loadProgressSeriesError: nil,
+            cloudState: .guest
+        )
+        defer { context.tearDown() }
+
+        await context.store.refreshReviewProgressBadgeIfNeeded(now: now)
+
+        XCTAssertNil(context.store.progressSnapshot)
+        XCTAssertEqual(
+            ReviewProgressBadgeState(
+                streakDays: 2,
+                hasReviewedToday: false,
+                streakFreezeAvailableCredits: 1,
+                streakFreezeCapacity: 2,
+                showsStreakFreezeBank: true,
+                isInteractive: true
+            ),
+            context.store.reviewProgressBadgeState
+        )
+        XCTAssertEqual(1, context.cloudSyncService.loadProgressSummaryCallCount)
+        XCTAssertEqual(0, context.cloudSyncService.loadProgressSeriesCallCount)
+    }
+
+    @MainActor
+    func testRefreshReviewProgressBadgeIfNeededLoadsBadgeSummaryWithoutBuildingSnapshot() async throws {
+        let database = try self.makeDatabase()
+        let workspace = try database.workspaceSettingsStore.loadWorkspace()
+        let cloudSettings = try database.workspaceSettingsStore.loadCloudSettings()
+
+        let now = try XCTUnwrap(parseIsoTimestamp(value: "2026-04-18T12:00:00.000Z"))
+        let requestRange = try makeTestProgressRequestRange(
+            now: now,
+            timeZone: TimeZone.current,
+            dayCount: 140
+        )
+        let serverSeries = try makeTestProgressSeries(
+            requestRange: requestRange,
+            reviewCountsByDate: [
+                "2026-04-01": 2
+            ],
+            generatedAt: "2026-04-18T11:59:00.000Z"
+        )
+        let serverSummary = try makeTestProgressSummary(
+            timeZone: requestRange.timeZone,
+            reviewDates: [requestRange.to],
+            generatedAt: "2026-04-18T11:59:00.000Z"
+        )
+        let context = try self.makeProgressStoreContext(
+            database: database,
+            workspaceId: workspace.workspaceId,
+            installationId: cloudSettings.installationId,
+            serverSummary: serverSummary,
+            serverSeries: serverSeries,
+            loadProgressSummaryError: nil,
+            loadProgressSeriesError: nil,
+            cloudState: .guest
+        )
+        defer { context.tearDown() }
+
+        await context.store.refreshReviewProgressBadgeIfNeeded(now: now)
+
+        XCTAssertNil(context.store.progressSnapshot)
+        XCTAssertEqual(
+            ReviewProgressBadgeState(
+                streakDays: 1,
+                hasReviewedToday: true,
+                streakFreezeAvailableCredits: 2,
+                streakFreezeCapacity: 2,
+                showsStreakFreezeBank: true,
+                isInteractive: true
+            ),
+            context.store.reviewProgressBadgeState
+        )
+        XCTAssertEqual(1, context.cloudSyncService.loadProgressSummaryCallCount)
+        XCTAssertEqual(0, context.cloudSyncService.loadProgressSeriesCallCount)
+    }
+}
